examples/restaurantProject/requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
asttokens==2.4.1
decorator==5.1.1
Django==4.2
django-nested-admin==4.0.2
exceptiongroup==1.2.0
executing==2.0.1
graphene==3.3
graphene-django==3.2.0
# Editable Git install with no remote (graphene-django-cruddals==1.0.0)
-e /home/juanjcardona13/Projects/Personal/CRUDDALS/Back/Python/graphene_django_cruddals
graphql-core==3.2.3
graphql-relay==3.2.0
ipython==8.22.2
jedi==0.19.1
matplotlib-inline==0.1.6
mypy==1.9.0
mypy-extensions==1.0.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
promise==2.3
prompt-toolkit==3.0.43
psycopg==3.1.18
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-monkey-business==1.0.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
text-unidecode==1.3
tomli==2.0.1
traitlets==5.14.2
typing_extensions==4.10.0
wcwidth==0.2.13
